presentation-listener/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
frozenlist==1.6.0
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
multidict==6.4.3
numpy==2.2.5
openai==1.76.0
portalocker==2.10.1
propcache==0.3.1
protobuf==6.30.2
PyAudio==0.2.14
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pywin32==310
qdrant-client==1.14.2
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==10.1
wsproto==1.2.0
yarl==1.20.0